refactor: 이미지 이름 변경
