+++ src/presentation/common/icons/asset/car-image-2.png
