+++ src/presentation/common/icons/asset/map-flag.png
